Remove build system, add search for remedy file, remove all in one
Cleaning up the plugin a bit, seems like going the build system
route was not worth it. Minimal gains, a lot of additional mental overhead.
As much as I love sublime the plugin api is not great for more complicated
things.

Instead by default plugin seeks for a nearby remedy file, hopefully this
will help a bit.